raspi_lora/requirements.txt
pyserial>=3.5
pynmea2>=1.4.0
smbus2>=0.4.0
